review fixes to athandler: changing int16/8 to int's and some minor issues.

+++ features/cellular/framework/API/CellularNetwork.h
@@ -372,7 +372,7 @@ public:
     /** Get the last 3GPP error code
      *  @return see 3GPP TS 27.007 error codes
      */
-    virtual uint8_t get_3gpp_error() = 0;
+    virtual int get_3gpp_error() = 0;
 
     /** Get the operator params
      *
